OBG-643. Restructure configuration properties
