+++ fintech-examples/fintech-impl/src/main/java/de/adorsys/opba/fintech/impl/config/FeignConfig.java
@@ -1,7 +1,7 @@
 package de.adorsys.opba.fintech.impl.config;
 
-import de.adorsys.opba.api.security.domain.DataToSign;
-import de.adorsys.opba.api.security.service.RequestSigningService;
+import de.adorsys.opba.api.security.external.domain.DataToSign;
+import de.adorsys.opba.api.security.external.service.RequestSigningService;
 import de.adorsys.opba.fintech.impl.properties.TppProperties;
 import feign.Request;
 import feign.RequestInterceptor;
@@ -13,11 +13,11 @@ import org.springframework.context.annotation.Configuration;
 import java.time.Instant;
 import java.util.UUID;
 
-import static de.adorsys.opba.fintech.impl.tppclients.HeaderFields.X_REQUEST_SIGNATURE;
-import static de.adorsys.opba.fintech.impl.tppclients.HeaderFields.FINTECH_ID;
-import static de.adorsys.opba.fintech.impl.tppclients.HeaderFields.X_TIMESTAMP_UTC;
-import static de.adorsys.opba.fintech.impl.tppclients.HeaderFields.X_REQUEST_ID;
 import static de.adorsys.opba.fintech.impl.tppclients.HeaderFields.COMPUTE_PSU_IP_ADDRESS;
+import static de.adorsys.opba.fintech.impl.tppclients.HeaderFields.FINTECH_ID;
+import static de.adorsys.opba.fintech.impl.tppclients.HeaderFields.X_REQUEST_ID;
+import static de.adorsys.opba.fintech.impl.tppclients.HeaderFields.X_REQUEST_SIGNATURE;
+import static de.adorsys.opba.fintech.impl.tppclients.HeaderFields.X_TIMESTAMP_UTC;
 
 /**
  * This class enhances requests, so that the include PSU IP address headers. This header should either become
