+++ opba-api-security/src/main/java/de/adorsys/opba/api/security/external/service/RequestSigningService.java
@@ -1,0 +1,8 @@
+package de.adorsys.opba.api.security.external.service;
+
+import de.adorsys.opba.api.security.external.domain.DataToSign;
+
+public interface RequestSigningService {
+
+    String signature(DataToSign dataToSign);
+}
