+++ opba-api-security/src/main/java/de/adorsys/opba/api/security/internal/config/ConfigConst.java
@@ -1,4 +1,4 @@
-package de.adorsys.opba.api.security.config;
+package de.adorsys.opba.api.security.internal.config;
 
 import lombok.experimental.UtilityClass;
