+++ opba-api-security/src/main/java/de/adorsys/opba/api/security/internal/service/RequestVerifyingService.java
@@ -1,6 +1,6 @@
-package de.adorsys.opba.api.security.service;
+package de.adorsys.opba.api.security.internal.service;
 
-import de.adorsys.opba.api.security.domain.DataToSign;
+import de.adorsys.opba.api.security.external.domain.DataToSign;
 
 public interface RequestVerifyingService {
     boolean verify(String signature, String encodedPublicKey, DataToSign dataToSign);
