--- opba-api-security/src/main/java/de/adorsys/opba/api/security/service/RequestSigningService.java
@@ -1,7 +1,0 @@
-package de.adorsys.opba.api.security.service;
-
-import de.adorsys.opba.api.security.domain.DataToSign;
-
-public interface RequestSigningService {
-    String signature(DataToSign dataToSign);
-}
