+++ opba-auth-rest-impl/src/main/java/de/adorsys/opba/tppauthapi/config/CookieProperties.java
@@ -5,18 +5,24 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.validation.annotation.Validated;
 
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
 import java.time.Duration;
+
+import static de.adorsys.opba.tppbankingapi.config.ConfigConst.API_CONFIG_PREFIX;
 
 @Data
 @Validated
 @Configuration
-@ConfigurationProperties(prefix = "cookie")
+@ConfigurationProperties(prefix = API_CONFIG_PREFIX + "cookie")
 public class CookieProperties {
-    private static final Integer MAX_AGE_DEFAULT = 300;
-
     private boolean secure = true;
-    private Duration maxAge = Duration.ofSeconds(MAX_AGE_DEFAULT);
     private boolean httpOnly = true;
     private String path = "/";
+
+    @NotNull
+    private Duration maxAge;
+
+    @NotBlank
     private String sameSite;
 }
