+++ opba-auth-rest-impl/src/main/java/de/adorsys/opba/tppauthapi/config/TppPsuAuthConfig.java
@@ -1,6 +1,6 @@
 package de.adorsys.opba.tppauthapi.config;
 
-import de.adorsys.opba.api.security.EnableTokenBasedApiSecurity;
+import de.adorsys.opba.api.security.internal.EnableTokenBasedApiSecurity;
 import de.adorsys.opba.tppauthapi.controller.TppAuthResponseCookieTemplate;
 import org.springframework.beans.factory.config.BeanDefinition;
 import org.springframework.context.annotation.Bean;
