+++ opba-banking-rest-api/src/main/java/de/adorsys/opba/tppbankingapi/config/ConfigConst.java
@@ -12,5 +12,10 @@ public class ConfigConst {
     /**
      * REST API {@code org.springframework.boot.context.properties.ConfigurationProperties} prefix.
      */
-    public static final String BANKING_API_CONFIG_PREFIX = "api.banking.";
+    public static final String API_CONFIG_PREFIX = "api.";
+
+    /**
+     * BANKING REST API {@code org.springframework.boot.context.properties.ConfigurationProperties} prefix.
+     */
+    public static final String BANKING_API_CONFIG_PREFIX = API_CONFIG_PREFIX + "banking.";
 }
