+++ opba-banking-rest-impl/src/test/java/de/adorsys/opba/tppbankingapi/controller/TestBankSearchPerformance.java
@@ -1,7 +1,7 @@
 package de.adorsys.opba.tppbankingapi.controller;
 
-import de.adorsys.opba.api.security.domain.DataToSign;
-import de.adorsys.opba.api.security.service.RequestSigningService;
+import de.adorsys.opba.api.security.external.domain.DataToSign;
+import de.adorsys.opba.api.security.external.service.RequestSigningService;
 import de.adorsys.opba.tppbankingapi.BaseMockitoTest;
 import de.adorsys.opba.tppbankingapi.dto.TestResult;
 import de.adorsys.opba.tppbankingapi.services.StatisticService;
@@ -27,10 +27,10 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
-import static de.adorsys.opba.api.security.domain.HttpHeaders.FINTECH_ID;
-import static de.adorsys.opba.api.security.domain.HttpHeaders.X_REQUEST_ID;
-import static de.adorsys.opba.api.security.domain.HttpHeaders.X_REQUEST_SIGNATURE;
-import static de.adorsys.opba.api.security.domain.HttpHeaders.X_TIMESTAMP_UTC;
+import static de.adorsys.opba.api.security.external.domain.HttpHeaders.FINTECH_ID;
+import static de.adorsys.opba.api.security.external.domain.HttpHeaders.X_REQUEST_ID;
+import static de.adorsys.opba.api.security.external.domain.HttpHeaders.X_REQUEST_SIGNATURE;
+import static de.adorsys.opba.api.security.external.domain.HttpHeaders.X_TIMESTAMP_UTC;
 import static de.adorsys.opba.tppbankingapi.TestProfiles.ONE_TIME_POSTGRES_ON_DISK;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
