+++ opba-embedded-starter-tests/src/test/java/de/adorsys/opba/starter/BasicOpenBankingStartupTest.java
@@ -1,6 +1,6 @@
 package de.adorsys.opba.starter;
 
-import de.adorsys.opba.api.security.service.RequestSigningService;
+import de.adorsys.opba.api.security.external.service.RequestSigningService;
 import de.adorsys.opba.protocol.xs2a.entrypoint.ais.Xs2aListAccountsEntrypoint;
 import de.adorsys.opba.protocol.xs2a.entrypoint.ais.Xs2aSandboxListTransactionsEntrypoint;
 import de.adorsys.opba.starter.config.FintechRequestSigningTestConfig;
