+++ opba-protocols/xs2a-protocol-tests/xs2a-bdd-wiremock/src/main/java/de/adorsys/opba/protocol/xs2a/tests/e2e/wiremock/mocks/config/FintechRequestSigningTestConfig.java
@@ -1,12 +1,12 @@
 package de.adorsys.opba.protocol.xs2a.tests.e2e.wiremock.mocks.config;
 
-import de.adorsys.opba.api.security.EnableSignRequestBasedApiSecurity;
-import de.adorsys.opba.api.security.config.RequestSigningConfig;
+import de.adorsys.opba.api.security.external.EnableOpenBankingRequestSigning;
+import de.adorsys.opba.api.security.external.config.RequestSigningConfig;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.validation.annotation.Validated;
 
 
-@EnableSignRequestBasedApiSecurity
+@EnableOpenBankingRequestSigning
 @ConfigurationProperties("security")
 @Validated
 public class FintechRequestSigningTestConfig extends RequestSigningConfig {
